test: add Z-Wave JS provider E2E tests, split into test modules
Add E2E tests that exercise the full LCM async_setup_entry lifecycle
with a real Z-Wave JS integration underneath. Split the monolithic
test_zwave_js.py into three focused modules: test_provider.py (provider
unit tests), test_events.py (push updates, notifications, duplicate
code handling), and test_e2e.py (full lifecycle integration tests).

Entire-Checkpoint: 8c1e2501d87e
